+++ src/main/java/com/codigo/monefy/users/persistence/entity/PasswordResetToken.java
@@ -14,7 +14,7 @@ public class PasswordResetToken {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     // Evitar dejar por defecto el fetch en Eager porque vas a evitar el problema N+1
     // N+1 cuando usas Eager cada PasswordResetToken va a arrastrar un SELECT adicional para traer el User, aun cuando no lo uses.
@@ -30,7 +30,7 @@ public class PasswordResetToken {
 
     public PasswordResetToken() {}
 
-    public PasswordResetToken(Long id, User user, String token, LocalDateTime expiryDate, boolean isUsed) {
+    public PasswordResetToken(Integer id, User user, String token, LocalDateTime expiryDate, boolean isUsed) {
         this.id = id;
         this.user = user;
         this.token = token;
